requirements.txt
numpy==1.24.2
pandas==1.5.3
pipenv==2023.2.18
scikit-learn==1.1.1
scipy==1.10.1
torch==1.13.1
seaborn
matplotlib
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.4.0
distlib==0.3.6
distro==1.8.0
dpath==2.1.5
dulwich==0.21.3
dvc==2.51.0
dvc-data==0.44.1
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.3.1
dvc-studio-client==0.6.1
dvc-task==0.2.0
filelock==3.10.7
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
hydra-core==1.3.2
idna==3.4
iterative-telemetry==0.0.8
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
nanotime==0.5.2
networkx==3.0
numpy==1.24.2
omegaconf==2.3.0
orjson==3.8.8
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
rich==13.3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.1
scipy==1.10.1
scmrepo==0.1.17
seaborn==0.12.2
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sqltrie==0.3.0
tabulate==0.9.0
threadpoolctl==3.1.0
tomlkit==0.11.7
torch==1.13.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
voluptuous==0.13.1
wcwidth==0.2.6
yarl==1.8.2
zc.lockfile==3.0.post1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.21.0
notebook==6.5.3
papermill==2.4.0
parso==0.8.3
textblob==0.15.3
spacy==3.5.1
spacytextblob==4.0.0
spacy-transformers
gensim
shap
xgboost==1.7.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
